Fix empty children regression (#21099)
## Objective

#18865 introduced a regression for an empty invocation of `children!`
and `related!`. The `recursive_spawn` macro has no case for zero inputs,
so the compiler rejects `children![]`. Whether this is particularly
meaningful isn't important; there are a number of situations in which
you might want your code to compile even with an empty set of children.

## Solution

A case has been added to `recursive_spawn` for no inputs, returning a
unit. This matches the previous behavior for `children!` and `related!`
exactly.